+++ CHANGELOG.rst
@@ -1,7 +1,11 @@
 1.0.24 (unreleased)
 -------------------
 
-- Nothing changed yet.
+- Use latest guillotina base classes
+  [vangheem]
+
+- Use field context if set
+  [vangheem]
 
 
 1.0.23 (2017-10-02)